+++ app/(auth)/forgot_password.tsx
@@ -1,4 +1,4 @@
-import ForgotPasswordScreen from '@/src/screens/auth/login/forgot_password_screen';
+import ForgotPasswordScreen from '@/src/screens/auth/forgot_password';
 
 export default function ForgotPasswordRoute() {
   return <ForgotPasswordScreen onBackToLogin={() => {}} />;
